# [client-v2] Migrate to use InputStream do read data