--- a/client-v2/src/main/java/com/clickhouse/client/api/data_formats/RowBinaryWithNamesFormatReader.java
+++ b/client-v2/src/main/java/com/clickhouse/client/api/data_formats/RowBinaryWithNamesFormatReader.java
@@ -1,12 +1,11 @@
 package com.clickhouse.client.api.data_formats;
 
-import com.clickhouse.client.api.ClientException;
 import com.clickhouse.client.api.data_formats.internal.AbstractBinaryFormatReader;
+import com.clickhouse.client.api.data_formats.internal.BinaryStreamReader;
 import com.clickhouse.client.api.metadata.TableSchema;
 import com.clickhouse.client.api.query.QuerySettings;
 import com.clickhouse.data.ClickHouseColumn;
 
-import java.io.EOFException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
@@ -30,9 +29,9 @@ public class RowBinaryWithNamesFormatReader extends AbstractBinaryFormatReader {
     public void readRecord(Map<String, Object> record) throws IOException {
         if (columns == null) {
             columns = new ArrayList<>();
-            int nCol = chInputStream.readVarInt();
+            int nCol = BinaryStreamReader.readVarInt(input);
             for (int i = 0; i < nCol; i++) {
-                columns.add(chInputStream.readUnicodeString());
+                columns.add(BinaryStreamReader.readString(input));
             }
 
             columns = Collections.unmodifiableList(columns);