… client-v2/src/main/java/com/clickhouse/client/api/query/QuerySettings.java
… client-v2/src/test/java/com/clickhouse/client/query/QueryTests.java