## Summary
Previous implementation was using ClickHouseInputStream and BinaryStreamUtils to read data. 
This PR removes any references for this two classes from client v2 code 

Bonus: 
There is a fix for premature summary. Operations like `queryAll` or `queryRecords` require to have result rows count. However summery doesn't guarantee to have it correct if no `wait_end_of_query=true` is set.