@chernser 2024-07-12
Just for curiosity I've benchmarked simple `select 1` executed with old client internals and new 
```
// old impl

Iteration   1: 0.003 ±(99.9%) 0.001 s/op
                 p0.00:   0.002 s/op
                 p0.50:   0.002 s/op
                 p0.90:   0.003 s/op
                 p0.95:   0.003 s/op
                 p0.99:   0.005 s/op
                 p0.999:  0.172 s/op
                 p0.9999: 0.192 s/op
                 p1.00:   0.192 s/op

Iteration   2: 0.004 ±(99.9%) 0.001 s/op
                 p0.00:   0.002 s/op
                 p0.50:   0.003 s/op
                 p0.90:   0.004 s/op
                 p0.95:   0.006 s/op
                 p0.99:   0.021 s/op
                 p0.999:  0.224 s/op
                 p0.9999: 0.238 s/op
                 p1.00:   0.238 s/op


// new impl



Iteration   1: 0.002 ±(99.9%) 0.001 s/op
                 p0.00:   0.001 s/op
                 p0.50:   0.001 s/op
                 p0.90:   0.002 s/op
                 p0.95:   0.002 s/op
                 p0.99:   0.004 s/op
                 p0.999:  0.127 s/op
                 p0.9999: 0.196 s/op
                 p1.00:   0.196 s/op

Iteration   2: 0.002 ±(99.9%) 0.001 s/op
                 p0.00:   0.001 s/op
                 p0.50:   0.001 s/op
                 p0.90:   0.002 s/op
                 p0.95:   0.002 s/op
                 p0.99:   0.004 s/op
                 p0.999:  0.089 s/op
                 p0.9999: 0.132 s/op
                 p1.00:   0.132 s/op
```